penumbra(storage): move `VersionedKeyhash` to `store` crate
