+++ crates/storage/src/store/substore.rs
@@ -12,7 +12,7 @@ use jmt::{
 use rocksdb::{ColumnFamily, IteratorMode, ReadOptions};
 use tracing::Span;
 
-use crate::{snapshot::RocksDbSnapshot, storage::VersionedKeyHash, Cache};
+use crate::{snapshot::RocksDbSnapshot, Cache};
 
 use jmt::storage::TreeWriter;
 
@@ -441,5 +441,45 @@ impl DbNodeKey {
         let node_key_slice = bytes.as_ref()[8..].to_vec();
         let node_key = borsh::BorshDeserialize::try_from_slice(&node_key_slice)?;
         Ok(DbNodeKey(node_key))
+    }
+}
+
+/// Represent a JMT key hash at a specific `jmt::Version`
+/// This is used to index the JMT values in RocksDB.
+#[derive(Clone, Debug)]
+pub struct VersionedKeyHash {
+    pub key_hash: KeyHash,
+    pub version: jmt::Version,
+}
+
+impl VersionedKeyHash {
+    pub fn new(version: jmt::Version, key_hash: KeyHash) -> Self {
+        Self { version, key_hash }
+    }
+
+    pub fn encode(&self) -> Vec<u8> {
+        let mut buf: Vec<u8> = self.key_hash.0.to_vec();
+        buf.extend_from_slice(&self.version.to_be_bytes());
+        buf
+    }
+
+    pub fn _decode(buf: Vec<u8>) -> Result<Self> {
+        if buf.len() != 40 {
+            Err(anyhow::anyhow!(
+                "could not decode buffer into VersionedKey (invalid size)"
+            ))
+        } else {
+            let raw_key_hash: [u8; 32] = buf[0..32]
+                .try_into()
+                .expect("buffer is at least 40 bytes wide");
+            let key_hash = KeyHash(raw_key_hash);
+
+            let raw_version: [u8; 8] = buf[32..40]
+                .try_into()
+                .expect("buffer is at least 40 bytes wide");
+            let version: u64 = u64::from_be_bytes(raw_version);
+
+            Ok(VersionedKeyHash { version, key_hash })
+        }
     }
 }
